refactor(gsharp): apply MR review feedback + add review tracker
Round-1 MR review fixes (still scaffolding-only; no implementation):

- : drop wrong `gsplat/` path prefix in 0000_init.md line 14
  (workspace mirror updated in parallel).
- + : drop gsplat/losses_depth.py; fold the four stub
  signatures into gsplat/losses.py per @amaximo's split-by-impl
  precedent (losses.py + losses_fused.py). binocular_disparity_l1
  and pearson_depth_loss appended to the existing Depth-losses
  section; masked_l1 / masked_ssim under a new "Mask-aware photometric
  wrappers (G-SHARP v0.2)" section. tests/test_losses_depth.py kept;
  only its docstring re-points at gsplat.losses. gsplat/regularizers.py
  kept as drafted (losses act on (pred, gt) pairs; regularizers act on
  a single tensor). gsplat/__init__.py untouched (no re-export to
  remove).
- : rename examples/dynamic_trainer.py ->
  examples/dynamic_surgical_trainer.py per @shsolanki + @amaximo.
  Internal docstring + NotImplementedError message updated.
- + : no code change. `gsplat/contrib/` namespace and
  `gsplat/init_utils.py` placement kept as-drafted; replies queued on
  the GitLab threads.

Plan markdown, RST proposal, and surgical-tutorial doc all updated to
match (component map, mermaid nodes, code references, Quickstart bash).
`python -m py_compile gsplat/losses.py` clean; grep clean for
losses_depth/dynamic_trainer in active code/docs/plan.

Adds new infra under planning/dev_updates/:
- 0001_scaffolding_commit.md (catch-up entry for c2652bc).
- 0002_post_review_changes.md (this round).
- mr_tracker.md + per-thread detail files under mr_comments/ — an
  MD-based tracker for review comments, summary-per-reviewer with a
  shared index.

Stale (flagged for next pass):
- planning/gsharp_gsplat_plan.html still has losses_depth.py and
  dynamic_trainer.py strings; regenerate from MD before next external
  share.
- planning/gsharp_gsplat_plan.md line 123 has the same gsplat/docs/...
  prefix bug as in a spot CodeRabbit didn't flag; left alone to
  keep scope tight.

Reauthored by: Rodolfo Lima <rlima@nvidia.com>
