+++ examples/dynamic_surgical_trainer.py
@@ -7,7 +7,7 @@ Wires together:
 - :class:`examples.datasets.endonerf.EndoNeRFParser` +
   :class:`EndoNeRFDataset`
 - :func:`gsplat.init_utils.multi_frame_depth_unprojection`
-- :func:`gsplat.losses_depth.binocular_disparity_l1` /
+- :func:`gsplat.losses.binocular_disparity_l1` /
   :func:`pearson_depth_loss` / :func:`masked_l1`
 - :func:`gsplat.regularizers.compute_tv_loss_targeted`
 - :class:`gsplat.training.TwoStageScheduler`
@@ -22,7 +22,9 @@ from __future__ import annotations
 
 
 def main() -> None:
-    raise NotImplementedError("vnath_gsharp: examples/dynamic_trainer.py pending")
+    raise NotImplementedError(
+        "vnath_gsharp: examples/dynamic_surgical_trainer.py pending"
+    )
 
 
 if __name__ == "__main__":
